Extract common util for quota adapters.
